% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_perc.R
\name{paste_perc}
\alias{paste_perc}
\title{Convert decimal value to a percent to be printed}
\usage{
paste_perc(x, digits = 0)
}
\arguments{
\item{x}{double or vector of doubles}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}
}
\value{
Character vector of a percent value
}
\description{
Convert decimal value to a percent to be printed
}
\examples{
x = 0.988
paste_perc(x)

x = c(0.444444, 0.2222, 0.5555)
paste_perc(x, 1)
}
