% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-datasets-schools-list.R
\docType{data}
\name{schools}
\alias{schools}
\title{schools}
\format{
A data frame with 140 rows and 3 variables:
\describe{
\item{ SchoolCode }{ character }
\item{ SchoolName }{ character }
\item{ SchoolLevel }{ character }
}
}
\source{
Aspen
}
\usage{
schools
}
\description{
Table of School Codes and School Name/Type.
}
\keyword{datasets}
